Renaming custom hook file
