+++ twiggy-app/src/pages/clothes/ClothesEditor.tsx
@@ -13,7 +13,7 @@ import {
 import { drawCircleTorso, drawOvalTorso, drawStickTorso } from '../../shapes/torsos';
 import { drawRoundLegs, drawRoundSkinnyLegs, drawStickLegs } from '../../shapes/legs';
 import { useAppSelector } from 'hooks/react-redux-hooks';
-import { Editors, useIsEditorActive } from 'hooks/active-editor-hooks';
+import { Editors, useIsEditorActive } from 'hooks/activeEditorHooks';
 import '../../styles/overlay.css';
 import { useState } from 'react';
